requirements.txt
altgraph==0.17.2
art==5.7
camelot-py==0.9.0
cffi==1.15.1
chardet==5.0.0
charset-normalizer==2.1.0
click==8.1.3
colorama==0.4.5
cryptography==37.0.4
cycler==0.11.0
decorator==5.1.1
Elixir==0.7.1
et-xmlfile==1.1.0
fonttools==4.34.4
future==0.18.2
greenlet==1.1.2
Jinja2==3.1.2
kiwisolver==1.4.4
MarkupSafe==2.1.1
matplotlib==3.5.2
numpy==1.23.1
opencv-python==4.6.0.66
openpyxl==3.0.10
packaging==21.3
pandas==1.4.3
pbr==5.9.0
pdfminer.six==20220524
pefile==2022.5.30
Pillow==9.2.0
pycparser==2.21
pyinstaller==5.2
pyinstaller-hooks-contrib==2022.8
PyMuPDF==1.20.1
pyparsing==3.0.9
PyPDF2==2.6.0
python-dateutil==2.8.2
pytz==2022.1
pywin32-ctypes==0.2.0
six==1.16.0
SQLAlchemy==1.4.39
sqlalchemy-migrate==0.11.0
sqlparse==0.4.2
tabulate==0.8.10
Tempita==0.5.2
xlrd==0.7.1
xlwt==0.7.2
